pyproject.toml
[build-system]
requires = ["setuptools", "numpy>=1.8", "cython"]
build-backend = "setuptools.build_meta"

[project]
name = "powerfit-em"
description = "Rigid body fitting of high-resolution structures in low-resolution cryo-electron microscopy density maps"
authors = [{ name = "Gydo C.P. van Zundert", email = "g.c.p.vanzundert@uu.nl" }]
maintainers = [{ name = "BonvinLab" }, { email = "bonvinlab.support@uu.nl" }]
requires-python = ">=3.10"
dynamic = ["readme", "version"]
license = "Apache-2.0"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Intended Audience :: Science/Research", 
    "Topic :: Scientific/Engineering :: Bio-Informatics"
]
dependencies = [
    "numpy>=1.8",
    # pyfftw binary wheel is not available on wasm or on macOS <14 arm, so fallback to numpy for wasm and macOS
    "pyfftw>=0.12.0; platform_machine != 'wasm32' and sys_platform != 'darwin'",
    "scipy",
    "tqdm",
    "rich",
    "molviewspec>=1.6.0",
]

[project.optional-dependencies]
opencl = ["pyopencl", "siphash24", "pyvkfft"]
pocl = ["pyopencl[pocl]", "siphash24", "pyvkfft"]
cuda = ["cupy-cuda13x", "siphash24", "pyvkfft"]
dev = [
    "pytest",
    "pytest-cov",
    "ruff",
    "cython-lint",
]

[project.urls]
issues = "https://github.com/haddocking/powerfit"
homepage = "https://github.com/haddocking/powerfit"
source = "https://github.com/haddocking/powerfit"
tutorial = "https://www.bonvinlab.org/education/Others/"

[project.scripts]
powerfit = "powerfit_em.powerfit:main"

[tool.setuptools]
# ext_modules is still in setup.py as numpy.get_include() 
# and cythonize() can not be used here

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
powerfit_em = ["data/*.npy", "correlators/kernels.cl", "correlators/kernels.cu", "_powerfit.pyx"]

[tool.setuptools.dynamic]
readme = {file = "README.md", content-type = "text/markdown"}
version = {attr = "powerfit_em.__version__"}

[tool.cibuildwheel]
# Skip build that do not work or are not needed
skip = "cp314* cp310-musllinux* cp311-musllinux* cp312-musllinux* *universal2"
archs = ["auto64"]
test-command = "powerfit --help"

# For some reason the source wheel for pyfftw is used instead of the binary wheel
# so we install its compilation dependencies manually
# Outside cibuildwheel images it installs the binary wheel
[[tool.cibuildwheel.overrides]]
select = "cp*-manylinux*"
before-test = "yum install -y fftw-devel"

[[tool.cibuildwheel.overrides]]
select = "*-musllinux*"
before-all = "apk add fftw-dev"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
pydocstyle.convention = "google"

[tool.ruff.lint.isort]
known-first-party = ["powerfit_em"]
